+++ docs/changelog.rst
@@ -24,6 +24,7 @@ Changelog
 * Numeric input fields will now accept simple math expressions.
 * Fixed the Editor View struggling to grab keyboard focus sometimes.
 * Fixed an issue that caused a project folder to be deleted when the ESC key was pressed in the confirmation dialog.
+* Fixed the Editor View's size when Hi-DPI is enabled.
 * Fixed minor aesthetic issues.
 
 **Lua API**